requirements.txt
langchain==0.0.267
chainlit==0.7.700
httpx
httpx-auth
python-dotenv==1.0.0
openai==0.27.8
PyPDF2==3.0.1
tiktoken
